requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
langchain>=0.3.7
langchain-openai>=0.2.2
langchain-community>=0.3.7
langchain-text-splitters>=0.3.2
psycopg[binary]>=3.2.3
python-dotenv>=1.0.1
pgvector>=0.2.5
pypdf>=5.1.0
